Small refactor

+++ AtomicSend/Prepare.cs
@@ -9,10 +9,7 @@ namespace AtomicSend
         public static async Task Stage(string connectionString, string destination)
         {
             var client = new ManagementClient(connectionString);
-            if (await client.QueueExistsAsync(destination))
-            {
-                await client.DeleteQueueAsync(destination);
-            }
+            if (await client.QueueExistsAsync(destination)) await client.DeleteQueueAsync(destination);
             await client.CreateQueueAsync(destination);
             await client.CloseAsync();
         }
