+++ Topologies/Prepare.cs
@@ -6,9 +6,10 @@ namespace Topologies
 {
     public static class Prepare
     {
-        public static async Task Stage(string connectionString, string inputQueue, string topicName, string rushSubscription, string currencySubscription)
+        public static async Task Stage(string connectionString, string inputQueue, string topicName,
+            string rushSubscription, string currencySubscription)
         {
-            ManagementClient client = await Cleanup(connectionString, inputQueue, topicName, rushSubscription, currencySubscription);
+            var client = await Cleanup(connectionString, inputQueue, topicName, rushSubscription, currencySubscription);
 
             var subscriptionDescription = new SubscriptionDescription(topicName, rushSubscription)
             {
@@ -31,7 +32,7 @@ namespace Topologies
                 {
                     Label = "rush"
                 },
-                Action = null,
+                Action = null
             };
             await client.CreateRuleAsync(topicName, rushSubscription, ruleDescription);
 
@@ -39,45 +40,32 @@ namespace Topologies
             {
                 Name = "MessagesWithCurrencyCHF",
                 Filter = new SqlFilter("currency = 'CHF'"),
-                Action = new SqlRuleAction("SET currency = 'Złoty'"),
+                Action = new SqlRuleAction("SET currency = 'Złoty'")
             };
             await client.CreateRuleAsync(topicName, currencySubscription, ruleDescription);
 
             await client.CloseAsync();
         }
 
-        private static async Task<ManagementClient> Cleanup(string connectionString, string inputQueue, string topicName, string rushSubscription, string currencySubscription)
+        private static async Task<ManagementClient> Cleanup(string connectionString, string inputQueue,
+            string topicName, string rushSubscription, string currencySubscription)
         {
             var client = new ManagementClient(connectionString);
 
             if (await client.SubscriptionExistsAsync(topicName, rushSubscription))
-            {
                 await client.DeleteSubscriptionAsync(topicName, rushSubscription);
-            }
 
             if (await client.SubscriptionExistsAsync(topicName, currencySubscription))
-            {
                 await client.DeleteSubscriptionAsync(topicName, currencySubscription);
-            }
 
-            if (await client.TopicExistsAsync(topicName))
-            {
-                await client.DeleteTopicAsync(topicName);
-            }
+            if (await client.TopicExistsAsync(topicName)) await client.DeleteTopicAsync(topicName);
 
-            var topicDescription = new TopicDescription(topicName)
-            {
-            };
+            var topicDescription = new TopicDescription(topicName);
             await client.CreateTopicAsync(topicDescription);
 
-            if (await client.QueueExistsAsync(inputQueue))
-            {
-                await client.DeleteQueueAsync(inputQueue);
-            }
+            if (await client.QueueExistsAsync(inputQueue)) await client.DeleteQueueAsync(inputQueue);
 
-            var queueDescription = new QueueDescription(inputQueue)
-            {
-            };
+            var queueDescription = new QueueDescription(inputQueue);
             await client.CreateQueueAsync(queueDescription);
             return client;
         }
